Fix the `puts()` vs `fastprint()` iss properly (see 53fe4a5)
Was a little too quick to the draw on the last commit.  `puts()` adds
EOL markers and does not flush by default.  I've added a temporary
`fastprint()` function that mimics the old behavior and have commented
on issue #151 asking to have this behavior rolled in to the main stream.

Revert "switch to use puts() instead of fastprint() (see 53fe4a5)"

This reverts commit 2b4cd5817badb58ecdaa263a17a224dcfb53f0b9.